[[ Bug 21923 ]] Fix memory leak when computing formattedWidth of buttons
This patch fixes a memory leak in the computation of the formattedWith
of buttons. In order to compute the formatted width, the label string
is split into an array of lines which are then measured. The storage
of the evaluated lines array has been changed to use an auto class to
ensure that it is not leaked.