Ensure no word "Standalone" is appended to the standalone name
The `revStandalonePreBuild` command prepends the word "Standalone"
to the standalone name, if there is already a file
`pFolder & "/" & <standalone_name>`. When building for both Linux
and Linux 64, the `revStandalonePreBuild` command is called twice,
but both times with `pFolder`.

In this case `pFolder` is `/Users/panos/Downloads/<standalone_folder>/Linux`.

So when `revStandalonePreBuild` is called for Linux 64, the file
`pFolder & "/" & <standalone_name>` already exists
(as it was created when building for Linux), so the word "Standalone"
was prepended to the standalone name.

The fix is to call `revStandalonePreBuild` with `tDirectory` instead, as
`tDirectory` is:

`/Users/panos/Downloads/<standalone_folder>/Linux/`
`/Users/panos/Downloads/<standalone_folder>/Linux x64/`

But we don't want the last /, so we actually need `char 1 to -2 of tDirectory`