[[ Bug 21926 ]] Fix memory leak when using message box put command
This patch fixes a memory leak which occurs when using the message
box targetting put command variant (i.e. put or put into msg). The
leak was occurring due to a failure to delete the handler name
and line number arguments (in the form of MCParmaeter instances). To
fix this, the MCParameter instances are now allocated on the stack
rather than on the heap.