[[ Bug 21787 ]] Fix memory leak when implicitly deleting substacks
This patch fixes a memory leak which occurs when a mainstack has
substacks and the mainstack is deleted.

When a stack is deleted it has all references to it removed and
then is scheduled for deletion (i.e. it is placed in the deleted
object pool for removal when the stack unwinds enough to ensure
there are no ptr references to it on the C stack).

If a stack has substacks then the substacks also have references
removed to them, including the references held by the mainstack.
This means that when the (main) stack is deleted from the deleted
object pool, it no longer has any references to its substacks.

This patch fixes the leak which occurs both when using the (very
context specific) destroystack() method in MCDispatcher and also
when using the general stack del() method.

In the first case it ensures that the substacks of the stack
passed to destroystack() are explicitly destroyed (using the
delete operator).

In the second case it ensures that scheduledelete() is called
for each substack after its del() method has been called.

The result is that regardless of how a stack comes to be
deleted, its MCStack object instance will be eventually
deleted (using the delete operator).