[[ Bug 21928 ]] Fix memory leak when using menu buttons
This patch fixes a memory leak when using menu buttons caused by
failing to delete the internal stack created to enable them to
function. The leak came about when object deletion was being
abstracted, and the code was being refactored to use object handles.
The button's menu pointer was changed to an object handle and
the explicit deletion which was there previously removed. As
object handles are weak references, this meant that they were
never being deleted.

The leak has been fixed by ensuring that the dodel() method can
be run on menus-as-stacks (which required an extra check in the
MCStack::removereferences method) and then additionally adding
an explicit delete operator call on the button's menu-as-stack
in MCButton::freemenu().