[[ Bug 21930 ]] Fix memory leak in use of GetVariableEx(UTF8) calls
This patch fixes a memory leak when the GetVariableEx(UTF8) external
V0 APIs are used. The leak is caused because these two functions return
a buffer which the caller is not responsible for freeing. The
functions must convert and create the internal MCStringRef value to
the format requested and the buffer must be freed at some point.
To fix the leak, the converted buffers which are returned are added
to the string release pool used elsewhere in the V0 external API,
and are freed when control returns to the engine from the external
handler.