[[ Bug 21931 ]] Fix leak when executing send or call command
This patch fixes a memory leak when executing a send or a call
command. The leak arose when an expression in the argument list
string successfully evaluated to a value. Rather than give the
resulting (exec) value to the parameter as an argument, it was
being set meaning it ended up being over-retained.