[[ Bug 21964 ]] Fix MCWidget::hittest for invisible/disabled widgets
This patch adds the same guard to `MCWidget::hittest` as seen in the `MCControl`
and `MCGroup` implementation where the hit test should return nil if the target
is not visible or the target is disabled and the current tool is browse.

This issue was causing `controlAtLoc` to return hidden widgets on all platforms
and touch events on mobile to be sent to hidden widgets.