[[ Bug 21928 ]] Fix memory leak when using menu buttons
This patch fixes a memory leak when using menu buttons caused by
failing to delete the internal stack created to enable them to
function. The leak came about when object deletion was being
abstracted, and the code was being refactored to use object handles.
The button's menu pointer was changed to an object handle and
the explicit deletion which was there previously removed. As
object handles are weak references, this meant that they were
never being deleted.

The leak has been fixed by ensuring that scheduledelete() is
called on the button menu stack after it has been removed from
the (menu related) environment in MCButton::freemenu. Additionally
deletion of the MCStack 'needs' member has been added to the
MCStack destructor.