[[ Bug 21959 ]] Fix memory leak when using clipboard functions
This patch corrects memory leaks in the clipboard functions where
they result in calling CopyData() or CopyTypeString(). Due to the
signatures of the methods being non-standard (i.e. returning a
+1 retain count rather than the standard bool return value and
out argument pattern) there are several places where the return
value has been passed to an auto class constructor. As the latter
retain their argument, this causes an over-retain.

The problem has been fixed by ensuring the return values of the
relevant methods are given to the autoclass using Give() rather
than passing to the constructor.