[[ Bug 21960 ]] Ensure private handler lookups are cached
This patch ensures that private handler calls only lookup the handler
the first time their MCComref or MCFuncref syntax node is executed.

To achieve this the handler lookup code has been moved from
MCKeywordsExecCommandOrFunction to MCKeywordsResolveCommandOrFunction.
The latter is now called directly from the MCComref or MCFuncref
execution methods as the cache is present in those nodes.