[[ Tests ]] Ensure clock extension loads in controlAtLoc tests
In the develop branch, loading an extension without previously
loading one of its dependencies will cause an error to be thrown.
Thus this test, which omitted the loading of the widgetutils
dependency, passed CI in the develop-9.0 branch due to subtleties
with setting of the result when loading extensions fails.

This patch resolves the issue by explicitly loading the clock
widget's required dependency.