[[ Bug 21987 ]] Fix pasteboard leak on macOS
This patch fixes a leak of an NSPasteboard on macOS due to not
using the 'releaseGlobally' method to release the engine's
selection clipboard before it is released in the normal manner
using 'release'. To fix the issue, the constructor of MCMacRawClipboard
now takes an (optional) second boolean parameter which indicates
whether it needs global release. This is then used by the factory
function for the system selection clipboard.