[[ Bug 21989 ]] Fix benign leak of apple events on startup on macOS
This patch fixes a benign leak of two AE descriptors on startup
on macOS. The leak was caused by duplicating the descriptors when
it was unnecessary to do so as the system had passed ownership to
the engine. Additionally this patch removes the unused 'process'
method from the MCPendingAppleEvent objective-c class.