[[ Bug 21992 ]] Fix memory leaks when sorting using MCSortnodes
This patch fixes memory leaks which can occur when using engine features
which perform sorting using MCSortnodes. Specifically, when sorting
fields, sorting cards of a stack, sorting the selection list for grouping
and sorting the selection list for cloning.

The leaks occur because of the presumption that MCSortnodes are POD types
and thus require no specific construction or destruction. However,
MCSortnodes fail to be POD as they have a destructor which releases any
MCValueRef sort key which they hold.

To fix the issue a new auto-class MCAutoArrayZeroedNonPod has been
added to foundation-auto. This is direct replication of MCAutoArray
but tailored to the case where the element type has a destructor and
that destructor does nothing if the element is zeroed. This auto-class
has then been used to replace previous uses of MCAutoArray in MCStack::sort,
MCField::sort and MCSellst::sort.

Additionally, the no longer used MCStringsExecSortOld function has been
removed.