[[ Bug 21996 ]] Fix memory leak when using filter elements of
This patch fixes a memory leak which occurs when using the elements
form of the filter command.

The leak was caused by failing to release converted element values
after use. It has been fixed by using an auto-class to hold the
string values being matched against in the wildcard and pattern
matching classes.