[[ Bug 21995 ]] Fix memory leaks when using sockets
This patch fixes three kinds of memory leak which can occur when
using sockets.

The first leak occurs if a socket suffered an error during its
use and has a non-nil error string attached. Failure to delete
the error member in the MCSocket destructor caused it to leak,
this has been corrected.

The second leak occurs during name->address resolution and is
of a weak object handle. The MCNToAMessageCallbackInfo struct ceased to
be POD when the change to using object handles was made. This has
been fixed by using MCMemoryCreate and MCMemoryDestroy to allocate
and deallocate the structs.

The third leak occurs when parsing the socket connection string when
it is composed of just a host name. In this case the input string
was being over-retained due to using assignment to an auto class
*and* MCValueRetain.