[[ Bug 22001 ]] Fix memory leak when using import eps
This patch fixes a memory leak which occurred when using the
import eps command. The leak was caused by a failure to close
an open IO_handle. The problem has been fixed by restructuring
the command implementation to ensure that the handle is always
closed if it was successfully opened.