[[ Bug 22005 ]] Fix memory leaks when using browser widget
This patch fixes two different types of leak which occur in the
browser widget.

The first is from the wrapper classes used to host the platform
specific callback interfaces for events and javascript handlers.
A pair of these wrapper class instances leak for every instance
of the browser widget on all platforms. The leak is caused by
not releasing the (reference-counted) instances correctly.

The second is from a failure to free the c-string form of the
current javascript handlers list. The last set will leak when
the browser widget instance is destroyed. This affects Mac,
Windows, Linux and iOS. The leak has been fixed by freeing the
c-string in the per-platform destructors.