[[ Bug 22010 ]] Fix memory leak when binding LCB foreign handlers
This patch fixes a memory leak which can occur when binding to LCB
foreign handlers. Previously, if a foreign handler was binding to
a C or Objective-C function and was present in an external code
library then a leak of the native code module handle wrapper could
occur.

This patch fixes the issue by holding any such handles in a per-module
array keyed by the name of the external native code library being loaded.