[[ Bug 21997 ]] Fix memory leak in MCStringSplitByDelimiter
This patch fixes a memory leak of the split values in the internal
function MCStringSplitByDelimiter.

The leak occurs because the split values are stored in an
MCAutoArray<MCValueRef> and then copied into a properlist, without
releasing the original values.

The leak has been fixed by switching to use an MCAutoStringRefArray
class and its TakeAsProperList method.