[[ Bug 22019 ]] Fix several memory leaks in the field object
This patch fixes four memory leaks caused by using field features.

The first leak which has been fixed occurred when a field object
had a field-level tabAlign setting. A failure to free the underlying
setting when the field was destroyed caused the leak.

The second leak which has been fixed occurred when fetching the
tabAlign property across multiple paragraphs. A failure to free
intermediate values used to compute whether the setting was mixed
or not caused the leak.

The third leak which has been fixed occurred when importing field
content with htmlText when the htmlText contained tabAlign settings.
A failure to free intermediate values used to compute the final
tabAlign settings caused the leak.

The fourth leak which has been fixed occurred when fetching the
styledText representation of the field or a chunk of the field. An
over-retain of the returned array caused the leak.