[[ Bug 21993 ]] Tweak __MCName and __MCString layouts
This patch tweaks both the __MCName and __MCString struct layouts
to ensure they are more efficient.

The __MCName struct has had the now unused 'hash' member removed
when compiling for 64-bit architectures as this is stored in
free bits elsewhere in the structure.

The __MCString struct has had its layout changed when compiling
for 32-bit architectures so that the numeric_value double is
8-byte aligned, a padding member to ensure the struct is 32-bytes
in size on all platforms has also been added.

Additionally a low-level memory test has been added which verifies
the sizes of all the core __MCValue derived structs are the
expected (hand calculated and optimized for) values.