[[ CI ]] Hacks to run tests through leak checking on macOS
This patch contains some hacks which enables running the test
suite through the macOS leaks tool.

The use of 'value pools' in libfoundation is disabled so that
the stacktrace information generated by the leak checker is useful
(that being said, this is only useful when using the Xcode memory
graph features - for just detecting whether there are leaks or not,
it makes no difference at all).

The closing of shared library handles which have been opened with
dlopen has been disabled so that leaks can be traced in externals
and thirdparty loaded library (again this is probably only germaine
when needing stacktraces to analyze, i.e. through Xcode).

Finally the testrunner has been modified to run each individual test
through leaks, with a leaks failure causing a test failure. There is
some attempt at mitigation of false-positives caused by some
pseudo-non-determinism in the leaks '--atExit' option. Sometimes
a test which does not leak will come back with 1000's of leaks all
over the place; presumably because something system related has
caused disruption of the at-exit leak checking process.