[[ Bug 22024 ]] Fix LCB's 'is among the elements of' operator
This patch fixes an issue with LCB's 'is among the elements of'
operator which causes it to return the wrong result when searching
for an array in another array.

The issue was caused by the internal __MCArrayIsEqualTo method
only using pointer equivalence to check for element equality
rather than MCValueIsEqualTo.

Additionally an explicit test for 'is among the elements of' has
been added to the array test file.