[[ Math ]] Standardardize domain and range error checking
This patch templatizes the math functions in order to standardize
the checks for domain and range errors in their computation. Any 
non-finite result from finite inputs is deemed an error - we use
floating point exception checking to distinguish domain errors 
(FE_INVALID and FE_DIVIDE_BY_ZERO) and range errors (FE_UNDERFLOW and
FE_OVERFLOW).

In many cases this was not previously being checked, hence a number
of new errors have been added to deal with the errors that may occur
for each math operation.

This patch also moves MCU_dofunc from utils to the exec-math module, 
and where it makes sense, splits off certain functions into their
own implementations. This prevents one value having to be recomputed
in each iteration of the values passed to `geometricMean`, as well as
removing the computation of the arithmetic mean for several functions
that don't use it.

Finally, this patch removes unnecessary division operations from the 
wrap and mod implementations. Although fmod does not appear to raise
the FE_INVALID exception when its second argument is zero, the result
is still non-finite and thus throw the correct error by defaulting to
the domain error in MCMathThrowFloatingPointException.