[[ Bug ]] Fix crash and behavior of popup widget
This patch resolves a crash of popup widget when called in the absence of
a valid mouse stack. `MCWidgetPopupAtLocationWithProperties` no longer
translates the location parameter to a global location based on the mouse
stack. All translation is done prior to the call.

- When called from LCB the point is now translated relative to the host
widget stack which is what it is documented to do.
- When called from LCS the translation is relative to the target. This
allows for non-mouse events to call `popup widget`. For example, tabbing
a text field into focus. If no location was specified in LCS then the
mouse location is used and therefore translation is relative to the mouse
stack. Execution errors are now thrown if it is not possible to obtain a
valid global location.