[[ Constants ]] Refactor constant lookup and value creation
This patch cleans up the implementation of constant lookup, replacing the
string and number value fields in the constant table's Cvalue struct with
a typed union. This enables existing string or number-based constants to
be specified as before, but constants with values that are MCValueRefs to
be represented in the table using specific values of the CvalueType enum.
The lookup mechanism is thereby unaffected. It also removes the MCConstant
syntax class, as it is no longer needed, and uses MCLiteral instead.

The constant values are interned so that there is only ever one instance
of the value in memory.

Previously the `true` and `false` constants were evaluated as strings. This
patch changes them to be evaluated as booleans. The only material difference
this should make is in use of the `is strictly` operator and when the constants
are elements of arrays passed to LCB handlers; in the latter case it is already
the LCB code's responsibility to handle variable typed in incoming array elements.

Furthermore the constant `infinity` has been added, evaluating to +inf (as a
real).