[[ PaletteActions ]] Fix potential exec error in palette actions
Previously the palette actions widget assumed the incoming "enabled"
key of an element of the action array and "checked" of a menu item
description were strings - this was only the case if either an
explicit string or one of the boolean constants (which are evaluated
as strings) were put into the array - if an expression had been used
(eg true is true) the widget would have thrown an error. This patch
fixes that possibility, which becomes an inevitability when the
boolean constants are changed to be evaluated as booleans.