[[ Bug ]] Fix failing LCB varargs test on Linux
This patch fixes the LCB varargs tests on Linux. The test was failing
due to the assumption that the %p printf format would render a nullptr
as 0x0 as it does on macOS - this was an incorrect assumption.

To resolve this, the varargs tests have been restructured, with the
nullptr test being separated out. Rather than check for a specific
output string, it just checks it begins with the known string (the
issue was a crash, so the only thing which needs to be checked is
that the crash doesn't happen!).