[[ Bug ]] Improve android mapFilePath for BFS
This patch improves android `mapFilePath` for BFS in the case where not all
architectures are built. Specifically for the case where `armv7` was not built.
Previously non-architecture specific files were found in the `armv7` build
folder. This patch will mean each the each build folder in turn will be checked
for existence and the first one found will be used.