[[ Cleanup ]] Implement MCEngineCurrentContextObject
This patch implement `MCEngineCurrentContextObject` to be reused where a function
needs a fallback default object. If the function is called by a widget script the
particular widget is the default object while if it is called by a library the
current card of the defaultStack is the default object.

This patch updates the engine module functions to use `MCEngineCurrentContextObject`
where appropriate.