[[ Bug ]] Fix handler function pointers on newer Android versions
This patch fixes an issue with the generation of handler function
pointers on newer versions of Android (starting with 9).

The issue is that libffi is failing to generate closure trampolines
in memory which has the executable permission bit set. This has been
worked around by explicitly changing the memory protection flags
to be read/write/exec *after* libffi has created the closure.