[[ Bug 22121 ]] Implement delete undo for widgets
This patch ensures that when widgets are added to the undo queue when the selected
objects are deleted that the undo will rebind the widget correctly. This special
case of deleting a widget but leaving it in a state where no data loss will occur if
it is replaced with undo is handled by a new method `delforundo`.

This method ensures that the widget on the undo queue is in its correct unbound state
with an updated rep if the widget was bound. When the undo operation is a replace of
a deleted widget then the widget is re-bound. When the undo operation is a delete of
a replaced widget then the `delforundo` method is again called.