[[ Bug 22104 ]] Throw error on objchunk of me if me can't be a parent
This patch ensures an error is thrown when a script refers to a child object
of me where me is a control that can not be a parent. For example,
`field "foo" of me` where `me` is a button. Previously, the resolved object
would just be `me` in this case the button.