[[ CEF ]] Fix crash on quit on linux
This patch resolves a crash when quitting while a card is open
with a browser widget on it. Firstly it moves `MCExtensionFinalize`
to after `MCstacks->closeall()` so that any extensions that need to
handle `OnClose` can do so and secondly it calls `MCscreen-DoRunloopActions()`
to ensure that anything requiring a runloop action to be performed
after `OnClose` won't be left hanging.