[[ TileCache ]] Fix tiles composited at wrong offset when group focus changed
This patch fixes an issue where changes to group keyboard focus which change the
effective rect of the group would not trigger an update of the layer rect,
leading to a mismatch between the previously cached tiles and the offset at
which they are rendered.

As groups don't render a focus border, the effective rect should not change when
focused.