[[ Bug 22760 ]] Fix 'external' handler invocation
This patch ensures that when a handler-ref is invoked from the main/ui thread
on mobile that a jump is made to the script/engine thread before being
executed.

After the handler is executed the JNI environment is reset for the calling
thread. This patch also fixes an issue where `MCJavaGetJObjectClassName` did
not call `MCJavaDoAttachCurrentThread` and therefore could access JNI
environment attached to a different thread.

(cherry picked from commit 15eff1554567cc63e5d6509b29201dc89d8ff20b)