[[ Bug 22130 ]] Disable deleted object pool creation whilst debugging
This patch adds a 'freeze' count to the deleted object pool logic. When the current
pool is frozen, no new pools will be created until it is thawed again. This means
that all objects that are deleted during that period are accumulated in the frozen
pool.

This ability is used inside MCB_prepmessage, which is the handler which causes the
engine to enter debug mode and send a message. Doing this prevents corruption of
the deleted object pools stack which appears to occur when the script editor
manipulates menus in the context of a debug-related message.

As the use of freezing the current object pool is only used in this one place it
has no effect on normal running code, only on code which runs in the context of
a debug-related message.