[21566] Fix crash when opening and closing modal windows
Problem:

When opening and closing modal windows continuously in a specific stack, at some point this error was thrown:

```
[NSApplication runModalSession:] may not be invoked inside of transaction begin/commit pair, or inside of transaction commit (usually this means it was invoked inside of a view's -drawRect: method.) The modal dialog has been suppressed to avoid deadlock.
```

After that, a crash occurred in a random place each time.

The previous error was thrown when this line was executed, in MCMacPlatformEndModalSession():

`[s_modal_sessions[t_final_index - 1] . window -> GetHandle() orderOut: nil];`

Solution:

It looks like in this case the suggested approach is to defer the operation until the next iteration of the main run loop:

https://forums.developer.apple.com/thread/74394
https://forums.developer.apple.com/thread/88825

So wrapping this line with:

`[performSelector: withObject: afterDelay:]`

causes the error to disappear, and I can no longer get the app to crash.