[[ Bug 22536 ]] Don't fail MCStringFindAndReplace if replacement not possible
This patch ensures that MCStringFindAndReplace won't fail if it determines
that the target string is native and the pattern string cannot ever be
native (and thus not exist in the target). The case this optimization occurs
is not hugely common as it requires the pattern string to have been 'checked'
at some point. The method will now just exit early in this case.