[[ Bug 22556 ]] Fix unpaired unlock of dragboard
This patch fixes an issue where `MCdragboard->Unlock()` was called when the
drag leaves the window when `MCdragboard->Lock()` had not been called. The
call to `Unlock()` has been changed to `PushUpdates()` so that the system
dragboard is updated with the current data so that it can be dropped onto
other applications etc. This change resolves an issue where after a number of
drags from LiveCode to another application drag and drop no longer appears to
function.