[[ Bug 18876 ]] Fix character duplication when deleting text on Android
This patch fixes an issue where using the backspace key to delete text
previously entered using a virtual keyboard on Android would result in extra
text being added to the field. This was due to both our own method overrides and
the underlying BaseInputConnection class sending synthesised KeyEvents to the
field. The implementation has been revised to prevent this.