[[ Bug 22577 ]] Ensure device token is returned in the correct form
For getting the device token, we were rendering an NSData into a formatted string,
which would use the debug representation. This worked fine until iOS 12.
On iOS 13, the debug representation of the NSData has changed, and does not
display the whole token. So we need to convert the NSData to a hex string
to get the correct representation.