[[ Bug 22614 ]] Resync mouse focus when control opacity changes
A change in a group or graphic's opacity can cause a change in the
control which should have the mouse focus - for example the
currently focused control may become transparent so that focus
transfers to any control underneath, or a control on a higher layer
than the focused control may become opaque and take the mouse focus.

This patch ensures that the mouse focus is resynced whenever any
group or graphic's opacity changes, so that a subsequent mouse event
has the correct target.