[[ Bug 22616 ]] Fix incorrect placement of native layers within moved group
This patch fixes bug 22616, where the native layer of a widget within a group
would be offset incorrectly from the parent layer when the group was moved.

This is due to position calculations being based on the rect of the parent
group, which was not updated until after the native layer of the widget was
repositioned, so the new position of the layer was based on the old position
of its container rather than the updated position.