[Bug 22482] Various fixes for widget Line Graph
Adds a decimal display of Y-axis labels. This prevents values close to one another to be labeled the same due to rounding errors.[18868], [22482], [19626]
Fixes a display bug if all Y-values are the same (NAN) [22482]
Fixes missing Y-values being displayed as 0, now it omits the values [18867]
Fixes missing first Y-values having the origin of 0,0 of widget
Allows to display incomplete Y-values.