[[ Bug 22607 ]] Implement system appearance support
This patch implements a `systemAppearance` property and a `systemAppearanceChanged`
message for macOS, Android and iOS.

The property returns if the application is running under either light or dark
mode. The message is sent to the current card of the defaultStack when the
system appearance changes (either automatically or when the user changes it).