[[ Bug 22605 ]] Use storyboard for iOS startup screens
This patch implements storyboard compilation and inclusion as the startup
and launch screens on iOS. The template storyboard allows for a single centered
and unscaled image and a background color which will be visible in transparent
areas of the image and areas outside the image rectangle.

Two variants of the compiled storyboard are required. The `LaunchScreen` is
loaded by the OS. The `StartupScreen` is loaded by the engine while it is
initializing. They are the same with the minor tweak of a custom class and
launch screen setting.