[22054] Ensure mobilePickPhoto() proceeds after taking a photo
On some Samsung devices, mobilePickPhoto() did not proceed after taking a photo
using the device camera. This happened because the temp file where the
photo would be stored until the user accepts it could not be created due to
insufficient permissions. This patch ensures the appropriate permissions are
set so as mobilePickPhoto() works as expected on all Android devices.