[[ libbrowser ]] Queue new load requests until the previous request is finished
This patch delays load requests sent to a WKWebView if there is on ongoing
request, waiting until that request has finished before making the new request.

This fixes an issue on iOS where the WebKit subprocess would be blocked from
reading HTML files within an app's sandboxed documents folder if multiple
requests are made without waiting for the previous request to begin.