[[ Android ]] Add utf-8 encoded c-string as java method return type
This patch allows Java method calls to receive a utf8-encoded c-string as a
return value.

The character 't' can currently be used in the signature string for a Java
method call to specify a utf-8 encoded c-string as an input parameter. This
character can now also be used to specify the type of the return value from
the method.