[[ Bug 21305 ]] Cache often used Win32 metrics
This patch caches the screen dpi and non-client metrics in the MCScreenDC
class on open, with them being updated whenever a WM_SETTINGCHANGE or
WM_DISPLAYCHANGE message is received. This metrics are used frequently,
especially when serializing fields, and caching vastly improves the speed
of the operations using them.

To facilitate the caching and update of the metrics, a new method
`updatemetrics()` has been added to MCScreenDC, which is called when opened
and during `processdesktopchanged()`.