[[ Bugfix 21305 ]] Add simple write buffer to Win32 file handle
This patch adds a very simple write buffer to the MCStdioFileHandle
implementation on Win32. The write buffer is used on non-pipe handles
and is initialized on first use of a write. The buffer has a fixed
size of 64k, and a maximum single write size of 16k. Buffering writes
causes a significant improvement in performance of stack serialization
as that process results in a lot of 1, 2 and 4 byte individual writes.

Additionally, a call to `MCS_flush` has been added after the end byte
is written when serializing stackfiles. This ensures that if any buffered
write data fails to be written, an error is generated and the original
stackfile is not deleted. This specific change affects all platforms.