[[ Bug 22688 ]] Don't set empty url on WKWebView
This patch fixes a bug where setting the url to empty would block subsequent
requests. Loading an empty url in the WKWebView browser would happen without
the expected 'decidePolicyForNavigationAction' message being sent to the
browser delegate as expected. To ensure the message is sent, we now load the
html text of a blank page into the browser instead.