[[ Bug 21387 ]] Provide means to create function ptr with abi
This patch adds a new low-level method `MCHandlerGetFunctionPtrWithAbi` which
generates a function ptr for a given handler ref with a specific callling
convention. This method is intended to be used by LCB modules which need to
create callbacks to pass to foreign handlers on 32-bit Win32 and thus might
require the non-default (cdecl) calling convention. On platforms other than
32-bit Win32, then method does the same as `MCHandlerGetFunctionPtr`.

A new global (libfoundation) define has been added `__HAS_MULTIPLE_ABIS__`
which is defined only for 32-bit Win32. This define is used to control the
addition of an `other_closures` member to the internal `__MCHandler` value
struct. This member is used to store a linked-list of closures (and function
pointers) made for a given `MCHandlerRef` instance.

The closure (and function pointer) for the default ABI, regardless of platform,
is still stored within the `__MCHandler` value struct so only those for non-
default ABIs are stored in the linked list.