[[ WindowsCI ]] Add editbin replacement script
This patch adds a python script which twiddles the appropriate bit
in a windows executable to make it a console application. Console 
applications do not open a separate console window when invoked on 
the command line - the command prompt only returns when the executable 
process finishes. This is the necessary behavior for the standalone
engine when running tests.