[[ Bug 22702 ]] Perform ad-hoc codesign of macOS standalones
This patch adds an ad-hoc code signature to macOS standalones after they have been
built. This is necessary because the unbound standalone engine is now signed so that
the IDE can be notarised. As binding a standalone invalidates the signature (as the
content of the executable changes), a replacement valid signature must be in place to
prevent errors when trying to run the resulting app on Catalina.