[[ Player ]] Reduce property changes performed when setting enabledTracks
This patch reduces the work done by platform players when setting the
enabledTracks property by only changing the enabled state of tracks
added or removed from the enabledTracks.

This is useful for some player implementations (e.g. Windows
MediaFoundation) which need to reload the media source when enabling or
disabling tracks.