[[ Emscripten ]] Ignore alt-key state when passing key events
This patch masks the alt key state for key events, so that characters
that require an alt+key combination to be entered can be input into
a field. Without doing so, such inputs would be interpreted as e.g.
alt+# rather than alt+3 (the combo that produces '#') or just '#' and
assumed to be a menu shortcut instead of a character to be added to
the field.