[[ Foundation ]] Prevent emission of logging parameters in release mode
This patch ensures that `MCAssert`, `MCLog` and `MCLogWithTrace` only generate
executable code when building in debug mode.

Previously, the arguments to these macros would still be emitted as source
resulting in unwanted (and unused) expression evaluation at runtime. This
was done to eliminate unused variable errors for variables only used as
part of arguments to those macros.

Now, the arguments are still emitted as source in release mode but they are
contained within an if which will never be taken at runtime. This will mean
that the code should also be eliminated at compile-time as well.