[[ emscripten ]] Keep hidden input field in fixed position in browser window
This patch changes the CSS style of the hidden input field used to
detect input events to ensure it is kept within the visible area of the
page within the browser. This change prevents the browser window
from scrolling when focus is passed to the hidden field, even in
browsers that don't support the 'preventScroll' passed to the
Element.focus method.