[[ Bug 22877 ]] Ensure mouse focus restored after relayer
This patch ensures if the source object of the relayer command has the mouse
focus it is restored correctly afterwards.

Previously the `CS_MFOCUSED` state of the card was being checked, however, this
is only ever set if `CS_MENU_ATTACHED`. This patch checks the `CS_MFOCUSED`
state of the source object instead.