[[ Bug 22935 ]] Defer closure of modal window until after modal session ends
This patch fixes an issue where closing a modal window would
immediately trigger a window event resulting in 'resumeStack' being
sent to the window behind. If a script called 'wait' while handling
that message then it would lead to an attempt to run the modal session
that had just been ended, causing an error.