[[ Build ]] Do not sign simulator .dylib externals
This patch removes signing of the simulator .dylib externals, since it
was causing a signing error in older simulator versions (10.2, 11.2, 12.1)
when trying to sign revsecurity when building with Xcode12.x. Note that
these .dylibs do not need to be signed at this point, since they are signed
during the notarization process anyway.

FYI the signing error was "no identity found", even though the same identity
worked fine in other externals (non-revsecurity) and in newer simulator versions.